Moved order formatting logic to reducer, fixed Component order source issue. Added another field to popup.
