+++ src/app/store/Order/Order.reducer.js
@@ -10,11 +10,39 @@
  */
 
 import { getIndexedProducts } from 'Util/Product';
+import { formatCurrency } from 'Util/Price';
 import {
     GET_ORDER_LIST,
     GET_ORDER,
     SET_ORDER_LOADING_STATUS
 } from './Order.action';
+
+const getFormattedDate = (rawDate = '') => {
+    const date = new Date(rawDate.replace(/\s/, 'T'));
+    const RADIX = 10;
+
+    const addLeadingZero = value => (value < RADIX ? `0${value}` : value);
+
+    const day = addLeadingZero(date.getDate());
+    const month = addLeadingZero(date.getMonth() + 1);
+
+    return `${day}.${month}.${date.getFullYear()}`;
+};
+
+const formatOrders = orders => orders.reduce((acc, order) => {
+    const { created_at, grand_total } = order;
+    const priceString = `${grand_total}${formatCurrency()}`;
+    const formattedDate = getFormattedDate(created_at);
+
+    return [
+        ...acc,
+        {
+            ...order,
+            grand_total: priceString,
+            created_at: formattedDate
+        }
+    ];
+}, []);
 
 export const initialState = {
     orderList: [],
@@ -33,10 +61,11 @@ const OrderReducer = (state = initialState, action) => {
     switch (type) {
     case GET_ORDER_LIST:
         const { items = [] } = orderList;
+        const formattedOrders = formatOrders(items);
 
         return {
             ...state,
-            orderList: items
+            orderList: formattedOrders
         };
 
     case GET_ORDER:
